# Optimise property counting for object validations